pyproject.toml
[project]
name = "image-gen-service"
version = "0.1.0"
description = "Local OpenAI-compatible image generation microservice for LoreWeave"
readme = "README.md"
requires-python = ">=3.12,<3.14"
license = { text = "MIT" }
authors = [{ name = "letuhao1994" }]

dependencies = [
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32,<0.33",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.6,<3",
    "aiosqlite>=0.20,<0.21",
    "structlog>=24.4,<25",
    "svix-ksuid>=0.6.2,<0.7",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27,<0.28",
    "asgi-lifespan>=2.1",     # lets the test client trigger FastAPI startup/shutdown hooks
    "ruff>=0.8",
]

# Forward-looking runtime deps (arch §14) deliberately deferred until the
# cycle that first needs them — adding them now pollutes the lockfile and
# slows Docker builds. Add per-cycle:
#   Cycle 3:  boto3, tenacity
#   Cycle 6:  respx (dev)
#   Cycle 9:  websockets, watchfiles

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "N",      # pep8-naming
    "S",      # flake8-bandit
    "ASYNC",  # flake8-async
    "RUF",    # ruff-specific
]
ignore = [
    "S101",   # asserts are fine; tests use them heavily
    "S104",   # binding 0.0.0.0 is correct in a container — uvicorn CMD does it intentionally
]
# Policy for other S-codes we'll hit in later cycles:
#   S113  httpx without timeout → address with explicit `timeout=` on every client
#   S311  non-crypto random     → use `secrets` module for IDs, `ksuid` for job ids
#   S608  string-formatted SQL  → use aiosqlite parameterised queries, never f-string SQL
# Use targeted `# noqa: Sxxx` only with a one-line rationale; don't blanket-ignore.

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S", "N"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: requires live external services (docker compose, GPU, MinIO) — skipped in unit runs",
]
